refactor: relocate graphics engine
